chore: create LoginPage/Language.ts to remove circular dependency
