+++ site/src/pages/LoginPage/Language.ts
@@ -1,0 +1,9 @@
+export const Language = {
+	emailLabel: "Email",
+	passwordLabel: "Password",
+	emailInvalid: "Please enter a valid email address.",
+	emailRequired: "Please enter an email address.",
+	passwordSignIn: "Sign In",
+	githubSignIn: "GitHub",
+	oidcSignIn: "OpenID Connect",
+};
