+++ site/src/pages/LoginPage/LoginPage.test.tsx
@@ -8,8 +8,8 @@ import {
 	waitForLoaderToBeRemoved,
 } from "testHelpers/renderHelpers";
 import { server } from "testHelpers/server";
+import { Language } from "./Language";
 import { LoginPage } from "./LoginPage";
-import { Language } from "./SignInForm";
 
 describe("LoginPage", () => {
 	beforeEach(() => {
