+++ site/src/pages/LoginPage/SignInForm.tsx
@@ -7,16 +7,6 @@ import { getApplicationName } from "utils/appearance";
 import { OAuthSignInForm } from "./OAuthSignInForm";
 import { PasswordSignInForm } from "./PasswordSignInForm";
 
-export const Language = {
-	emailLabel: "Email",
-	passwordLabel: "Password",
-	emailInvalid: "Please enter a valid email address.",
-	emailRequired: "Please enter an email address.",
-	passwordSignIn: "Sign In",
-	githubSignIn: "GitHub",
-	oidcSignIn: "OpenID Connect",
-};
-
 const styles = {
 	root: {
 		width: "100%",
